refactor: rename example
