+++ apps/paper-example/android/app/build.gradle
@@ -77,9 +77,9 @@ android {
     buildToolsVersion rootProject.ext.buildToolsVersion
     compileSdk rootProject.ext.compileSdkVersion
 
-    namespace "com.example"
+    namespace "com.paperexample"
     defaultConfig {
-        applicationId "com.example"
+        applicationId "com.paperexample"
         minSdkVersion rootProject.ext.minSdkVersion
         targetSdkVersion rootProject.ext.targetSdkVersion
         versionCode 1
