+++ apps/paper-example/android/app/src/main/java/com/paperexample/MainActivity.kt
@@ -1,6 +1,5 @@
-package com.example
+package com.paperexample
 
-import android.os.Bundle;
 import com.facebook.react.ReactActivity
 import com.facebook.react.ReactActivityDelegate
 import com.facebook.react.defaults.DefaultNewArchitectureEntryPoint.fabricEnabled
@@ -12,8 +11,8 @@ class MainActivity : ReactActivity() {
    * Returns the name of the main component registered from JavaScript. This is used to schedule
    * rendering of the component.
    */
-  override fun getMainComponentName(): String = "Example"
-  
+  override fun getMainComponentName(): String = "PaperExample"
+
   override fun onCreate(savedInstanceState: Bundle?) {
     super.onCreate(null)
   }
