+++ apps/paper-example/android/app/src/main/java/com/paperexample/MainApplication.kt
@@ -1,4 +1,4 @@
-package com.example
+package com.paperexample
 
 import android.app.Application
 import com.facebook.react.PackageList
