+++ apps/paper-example/android/app/src/main/res/values/strings.xml
@@ -1,3 +1,3 @@
 <resources>
-    <string name="app_name">Example</string>
+    <string name="app_name">PaperExample</string>
 </resources>
