+++ apps/paper-example/app.json
@@ -1,4 +1,4 @@
 {
-  "name": "Example",
-  "displayName": "Example"
+  "name": "PaperExample",
+  "displayName": "PaperExample"
 }
