+++ apps/paper-example/ios/PaperExample/AppDelegate.h
