+++ apps/paper-example/ios/PaperExample/Images.xcassets/AppIcon.appiconset/Contents.json
