+++ apps/paper-example/ios/PaperExample/Images.xcassets/Contents.json
