+++ apps/paper-example/ios/PaperExample/PrivacyInfo.xcprivacy
