+++ apps/paper-example/ios/PaperExampleTests/Info.plist
