+++ apps/paper-example/ios/PaperExampleTests/PaperExampleTests.m
@@ -7,11 +7,11 @@
 #define TIMEOUT_SECONDS 600
 #define TEXT_TO_LOOK_FOR @"Welcome to React"
 
-@interface ExampleTests : XCTestCase
+@interface PaperExampleTests : XCTestCase
 
 @end
 
-@implementation ExampleTests
+@implementation PaperExampleTests
 
 - (BOOL)findSubviewInView:(UIView *)view matching:(BOOL (^)(UIView *view))test
 {
