+++ apps/paper-example/ios/Podfile
@@ -16,7 +16,7 @@ if linkage != nil
   use_frameworks! :linkage => linkage.to_sym
 end
 
-target 'Example' do
+target 'PaperExample' do
   config = use_native_modules!
 
   use_react_native!(
@@ -25,7 +25,7 @@ target 'Example' do
     :app_path => "#{Pod::Config.instance.installation_root}/.."
   )
 
-  target 'ExampleTests' do
+  target 'PaperExampleTests' do
     inherit! :complete
     # Pods for testing
   end
